factored out graphql server helper
