+++ examples/graphql/server_graphql.py
@@ -1,20 +1,13 @@
 import asyncio
-import json
 import logging
 import sys
 from pathlib import Path
-from typing import AsyncGenerator, Tuple
 
-from graphql import build_schema, subscribe, parse
+from graphql import build_schema
 
-from rsocket.frame_helpers import str_to_bytes
-from rsocket.graphql.helpers import execute_query_in_payload, get_graphql_params
-from rsocket.helpers import create_future
-from rsocket.payload import Payload
-from rsocket.routing.request_router import RequestRouter
+from rsocket.graphql.server_helper import graphql_handler
 from rsocket.routing.routing_request_handler import RoutingRequestHandler
 from rsocket.rsocket_server import RSocketServer
-from rsocket.streams.stream_from_async_generator import StreamFromAsyncGenerator
 from rsocket.transports.tcp import TransportTCP
 
 
@@ -33,54 +26,15 @@ def greetings(*args):
     return results()
 
 
-class GraphqlRequestHandler:
+with (Path(__file__).parent / 'rsocket.graphqls').open() as fd:
+    schema = build_schema(fd.read())
 
-    def __init__(self):
-        with (Path(__file__).parent / 'rsocket.graphqls').open() as fd:
-            schema = build_schema(fd.read())
-
-        schema.query_type.fields['greeting'].resolve = greeting
-        schema.subscription_type.fields['greetings'].subscribe = greetings
-
-        router = RequestRouter()
-
-        @router.response('graphql')
-        async def graphql_query(payload: Payload):
-            execution_result = await execute_query_in_payload(payload, schema)
-
-            response_data = str_to_bytes(json.dumps({
-                'data': execution_result.data
-            }))
-
-            return create_future(Payload(response_data))
-
-        @router.stream('graphql')
-        async def graphql_subscription(payload: Payload):
-            async def generator() -> AsyncGenerator[Tuple[Payload, bool], None]:
-                data = json.loads(payload.data.decode('utf-8'))
-                params = get_graphql_params(data, {})
-                document = parse(params.query)
-
-                async for execution_result in await subscribe(
-                        schema,
-                        document,
-                        operation_name=params.operation_name
-                ):
-                    item = execution_result.data
-                    response_data = str_to_bytes(json.dumps({
-                        'data': item
-                    }))
-                    yield Payload(response_data), False
-
-                yield Payload(), True
-
-            return StreamFromAsyncGenerator(generator)
-
-        self.router = router
+schema.query_type.fields['greeting'].resolve = greeting
+schema.subscription_type.fields['greetings'].subscribe = greetings
 
 
 def handler_factory():
-    return RoutingRequestHandler(GraphqlRequestHandler().router)
+    return RoutingRequestHandler(graphql_handler(schema, 'graphql'))
 
 
 async def run_server(server_port):
